requirements.txt
allure-pytest==2.8.18
allure-python-commons==2.8.18
atomicwrites==1.4.0
attrs==20.2.0
colorama==0.4.4
importlib-metadata==2.0.0
iniconfig==1.1.1
packaging==20.4
pinyin==0.4.0
pluggy==0.13.1
py==1.9.0
pyautoit-win64==1.0.3
pyparsing==2.4.7
pytest==6.1.1
PyYAML==5.3.1
retrying==1.3.3
selenium==3.141.0
six==1.15.0
toml==0.10.1
urllib3==1.25.11
zipp==3.3.1
